refactor(types): divided the types.rs into academic.rs auth.rs and schedule.rs the types.rs was getting to big
