+++ src-tauri/src/types/auth.rs
@@ -1,0 +1,16 @@
+use std::collections::HashMap;
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct LoginResponse {
+    pub session_id: String,
+    pub aluno_ids: HashMap<String, String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct StudentInfo {
+    pub photo_data: Option<String>,
+    pub student_name: String,
+    pub course: String,
+}
